Pull syntax functions down into the relevant classes instead of relying on extensions.
