+++ orbit-core/src/commonMain/kotlin/org/orbitmvi/orbit/internal/LazyCreateContainerDecorator.kt
@@ -29,7 +29,7 @@ import kotlinx.coroutines.flow.flow
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerDecorator
 import org.orbitmvi.orbit.syntax.ContainerContext
-import org.orbitmvi.orbit.syntax.simple.intent
+import org.orbitmvi.orbit.syntax.intent
 
 public class LazyCreateContainerDecorator<STATE : Any, SIDE_EFFECT : Any>(
     override val actual: Container<STATE, SIDE_EFFECT>,
