+++ orbit-core/src/commonTest/kotlin/org/orbitmvi/orbit/internal/ContainerLifecycleTest.kt
@@ -25,7 +25,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
