+++ orbit-core/src/commonTest/kotlin/org/orbitmvi/orbit/internal/CoroutineNameTest.kt
@@ -6,8 +6,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
 import kotlin.coroutines.coroutineContext
 import kotlin.random.Random
 import kotlin.test.Test
