+++ orbit-core/src/commonTest/kotlin/org/orbitmvi/orbit/internal/SubIntentTest.kt
@@ -30,8 +30,6 @@ import kotlinx.coroutines.test.runTest
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.annotation.OrbitExperimental
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
-import org.orbitmvi.orbit.syntax.simple.subIntent
 import org.orbitmvi.orbit.test.test
 import kotlin.random.Random
 import kotlin.test.Test
