+++ orbit-core/src/commonTest/kotlin/org/orbitmvi/orbit/internal/repeatonsubscription/RepeatOnSubscriptionTest.kt
@@ -21,10 +21,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
-import org.orbitmvi.orbit.syntax.simple.reduce
-import org.orbitmvi.orbit.syntax.simple.repeatOnSubscription
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
