+++ orbit-test/src/commonTest/kotlin/org/orbitmvi/orbit/test/IntentJobsTest.kt
@@ -24,8 +24,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.reduce
 import kotlin.coroutines.coroutineContext
 import kotlin.test.Test
 import kotlin.test.assertFails
