+++ orbit-test/src/commonTest/kotlin/org/orbitmvi/orbit/test/RepeatOnSubscriptionTest.kt
@@ -20,9 +20,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.reduce
-import org.orbitmvi.orbit.syntax.simple.repeatOnSubscription
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
