+++ orbit-viewmodel/src/test/kotlin/org/orbitmvi/orbit/viewmodel/AndroidIdlingResourceTest.kt
@@ -37,7 +37,6 @@ import org.junit.runner.RunWith
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.intent
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import kotlin.test.AfterTest
