+++ samples/orbit-calculator/src/main/kotlin/org/orbitmvi/orbit/sample/calculator/CalculatorViewModel.kt
@@ -27,8 +27,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import kotlinx.parcelize.Parcelize
 import org.orbitmvi.orbit.ContainerHost
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.reduce
 import org.orbitmvi.orbit.viewmodel.container
 import java.math.BigDecimal
 import java.math.RoundingMode
