+++ samples/orbit-posts/src/main/kotlin/org/orbitmvi/orbit/sample/posts/app/features/postdetails/viewmodel/PostDetailsViewModel.kt
@@ -27,8 +27,6 @@ import org.orbitmvi.orbit.annotation.OrbitExperimental
 import org.orbitmvi.orbit.sample.posts.domain.repositories.PostOverview
 import org.orbitmvi.orbit.sample.posts.domain.repositories.PostRepository
 import org.orbitmvi.orbit.sample.posts.domain.repositories.Status
-import org.orbitmvi.orbit.syntax.simple.reduce
-import org.orbitmvi.orbit.syntax.simple.subIntent
 import org.orbitmvi.orbit.viewmodel.container
 
 class PostDetailsViewModel(
