+++ samples/orbit-text/src/main/kotlin/org/orbitmvi/orbit/sample/text/app/TextViewModel.kt
@@ -22,9 +22,6 @@ import kotlinx.coroutines.delay
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.container
-import org.orbitmvi.orbit.syntax.simple.blockingIntent
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.reduce
 import kotlin.random.Random
 
 @Suppress("MagicNumber")
